requirements.txt
pillow==12.1.1
requests==2.33.0
